agent: use authenicated request while waiting for the analyzer
If keystone is used as an authentication backend, the response to
the test request from the agent, to test if the analyzer is up,
will be 401.
This patch authenticate the agent before sending the test request,
so that it can be sent with authentication material.

Co-authored-by: Sylvain Baubeau <sbaubeau@redhat.com>
Change-Id: Ib2f3d585520a9c61ff469720edfb413fcc392400
Reviewed-on: http://softwarefactory-project.io/r/4944
Reviewed-by: Sylvain Afchain <safchain@gmail.com>
Tested-by: Sylvain Afchain <safchain@gmail.com>
Workflow: Sylvain Afchain <safchain@gmail.com>
Tested-by: Jenkins CI <jenkins@softwarefactory-project.io>