capture: return status error if not registered
Is register node failed on the server side due to
interface down for example the status returned was
OK. This patch fix that so that a capture will be
started when interface set to up.

Add query in the reply so that we can get capture
UUID and gremlin in logs.

Change-Id: If1ef4d8bb85dc8f845af0ab0116611e70407a16a