travis: create multiarch docker images
Create docker images after a successful build.
Because ppc64le didn't have a gcr.io/distroless/base
base image we've used Alpine for now.

sudo has been removed from the Makefile and its
assumed the user is in the docker group which
is sufficient to build (and run) docker images.

Docker images are created after success and now
include the date, travis build number and architecture
in the tag. These are combined in a subsequent stage
after the two architecture jobs are completed.

The docker latest tag is only added for tagged
releases consistent with other docker projects
that push incrementally. The latest is the default
pulled image so its more important to be stable.

The travis build has been changed to a container
build, and the apt plugin to install distro packages.

Because of the protoc version 3 requirement this
is built from source as a binary package didn't
exist for both architectures. We've used the
travis cache plugin to save a built version
(though currently ppc64le doesn't but this will
be fixed with travis).

Binaries now include the architecture as the
suffix.

Tagged releases are pushed to the github.